code optmized
