+++ pages/about.html
@@ -34,10 +34,10 @@
                 <aside class="sidebar">
                     <a class="new-stuff" href="#">Logout</a>
                     <a href="../index.html">Home</a>
-                    <a href="#">Program</a>
-                    <a href="#">Join</a>
-                    <a href="#">Sponsor</a>
-                    <a href="#">News</a>
+                    <a class="program" href="#">Program</a>
+                    <a class="join" href="#">Join</a>
+                    <a class="sponsor" href="#">Sponsor</a>
+                    <a class="news" href="#">News</a>
                     <a class="CC" href="#">CC Campaign</a>
                 </aside>
             </section>
